stanova_brm {stanova}	R Documentation

Estimate ANOVA-type models with brms

Description

Estimate ANOVA-type models with brms

Usage


stanova_brm(formula, data, model_fun, ..., check_contrasts = "contr.bayes")

Arguments

formula
a formula describing the statistical model to be estimated. Passed to the rstanarm::stan_ function specified in model_fun.

data
data.frame containing the data.

model_fun
character string identifying the rstanarm function that should be used for fitting (omitting the stan_ prefix) such as "glm" or "glmer".

...
further arguments passed to the rstanarm function used for fitting. Typical arguments are prior, prior_intercept, chain, iter, or core.

check_contrasts
character string (of length 1) denoting a contrast function or a contrast function which should be assigned to all character and factor variables in the model (as long as the specified contrast is not the global default). Default is contr.bayes. Set to NULL to disable the check.

Examples



## Not run: 
fit_warp_brms <- stanova_brm(breaks ~ wool * tension, data = warpbreaks,
                             chains = 2, iter = 1000)
summary(fit_warp_brms)

## End(Not run)

[Package stanova version 0.3-0 Index]