stanova_lmer {stanova}	R Documentation

Estimate ANOVA-type mixed models with rstanarm

Description

Estimate ANOVA-type mixed models with rstanarm

Usage

stanova_lmer(formula, data, check_contrasts = "contr.bayes", ...)

stanova_glmer(formula, data, family, check_contrasts = "contr.bayes", ...)

Arguments

formula
a formula describing the full mixed-model to be fitted. Passed to rstanarm::stan_g/lmer.

data
data.frame containing the data.

check_contrasts
character string (of length 1) denoting a contrast function which should be assigned to all character and factor variables in the model (as long as the specified contrast is not the global default). Default is contr.bayes. Set to NULL to disable the check.

...
further arguments passed to the rstanarm function used for fitting. Typical arguments are prior, prior_intercept, chain, iter, or core. Ensure that arguments that cannot be evaluated in the given context (e.g., weight arguments to binomial models in which only an unquoted column name is passed) are enclosed in quote().

family
family argument passed to stan_glmer.

Note

These functions are only wrappers around stanova setting model_fun to "glmer" (and family = "gaussian" for stanova_lmer).

Examples

data("Machines", package = "MEMSS")

m_machines <- stanova_lmer(score ~ Machine + (Machine|Worker),
                           data=Machines, chains = 2, iter = 500)
summary(m_machines) ## default: difference from intercept

summary(m_machines, diff_intercept = FALSE) ## alt: marginal means

out_array <- stanova_samples(m_machines)
str(out_array)

out_array2 <- stanova_samples(m_machines, dimension_chain = 2)
str(out_array2)

bayesplot::mcmc_trace(out_array2$`(Intercept)`)
bayesplot::mcmc_trace(out_array2$Machine)

out_df <- stanova_samples(m_machines, return = "data.frame")
str(out_df)

data("obk.long", package = "afex")

m2 <- stanova_lmer(value ~ treatment * phase + (1|id), obk.long,
                   chains = 2, iter = 500)

m2
summary(m2)

## with continuous variable
data(md_16.4, package = "afex")
md_16.4$cog <- scale(md_16.4$cog, scale=FALSE)

m_cont0 <- stanova_lmer(induct ~ cog + (cog|room:cond), md_16.4,
                        chains = 2, iter = 500)
summary(m_cont0)

# with interaction:
m_cont1 <- stanova_lmer(induct ~ cond*cog + (cog|room:cond), md_16.4,
                        chains = 2, iter = 500)
summary(m_cont1)

summary(m_cont1, diff_intercept = TRUE)

### glmer models

## binomial model
cbpp <- lme4::cbpp
cbpp$prob <- with(cbpp, incidence / size)
example_model <- stanova_glmer(prob ~ period + (1|herd),
                            data = cbpp, family = binomial,
                            weight = quote(size),
                            chains = 2, cores = 1, seed = 12345, iter = 500)
## Note: we need to quote() the weight column.
## Alterntively passing the column directly also works: cpbb$weights

summary(example_model)

## poisson model
data(Salamanders, package = "glmmTMB")
gm1 <- stanova_glmer(count~spp * mined + (1 | site), data = Salamanders,
                   family = "poisson",
                   chains = 2, cores = 1, seed = 12345, iter = 500)
summary(gm1)

[Package stanova version 0.1-0 Index]