stanova {stanova}	R Documentation

Estimate ANOVA-type models with rstanarm

Description

Estimate ANOVA-type models with rstanarm

Usage


stanova(formula, data, model_fun, ..., check_contrasts = "contr.bayes")

Arguments

formula
a formula describing the statistical model to be estimated. Passed to the rstanarm::stan_ function specified in model_fun.

data
data.frame containing the data.

model_fun
character string identifying the rstanarm function that should be used for fitting (omitting the stan_ prefix) such as "glm" or "lmer".

...
further arguments passed to the rstanarm function used for fitting. Typical arguments are prior, prior_intercept, chain, iter, or core. Ensure that arguments that cannot be evaluated in the given context (e.g., weight arguments to binomial models in which only an unquoted column name is passed) are enclosed in quote().

check_contrasts
character string (of length 1) denoting a contrast function which should be assigned to all character and factor variables in the model (as long as the specified contrast is not the global default). Default is contr.bayes. Set to NULL to disable the check.

Examples



fit_warp <- stanova(breaks ~ wool * tension, data = warpbreaks,
                    prior = rstanarm::student_t(3, 0, 3, autoscale = FALSE),
                    model_fun = "glm",
                    chains = 2, iter = 500)
summary(fit_warp)

####

data("Machines", package = "MEMSS")

## note: model_fun = "lmer" only works after library("rstanarm")
## but you can simply use model_fun = "glmer" with family = "gaussian"
m_machines <- stanova(score ~ Machine + (Machine|Worker),
                      model_fun = "glmer", family = "gaussian",
                      data=Machines, chains = 2, iter = 500)
summary(m_machines)

## Not run: 
## negative binomial
## requires attaching rstanarm
library("rstanarm")
fit6a <- stanova(Days ~ Sex/(Age + Eth*Lrn), data = MASS::quine,
                 model_fun = "glm.nb",
                 link = "log", prior_aux = exponential(1.5),
                 chains = 2, iter = 200) # for speed of example only
summary(fit6a)

## End(Not run)


[Package stanova version 0.1-0 Index]